12-RAG-python-markdown/requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Document Loaders
unstructured>=0.10.0  # For Markdown loading
unstructured[markdown]>=0.10.0  # Markdown support
markdown>=3.4.0  # Required dependency for unstructured markdown support

# Optional: For environment variable management
python-dotenv>=1.0.0
